Move default color and shape enums back to tile subpackage (still visible), fix internal and test imports
