--- src/tilematch_tools/model/tile_shape.py
@@ -1,12 +1,0 @@
-"""
-    :module_name: tile_shape
-    :module_summary: a class that dictates the shape of a given tile
-    :module_author: Nathan Mendoza (nathancm@uci.edu)
-"""
-
-from enum import IntEnum
-
-class TileShape(IntEnum):
-    SQUARE = 1
-    DIAMOND = 2
-    CIRCULAR = 3
