+++ src/tilematch_tools/model/tiles/__init__.py
@@ -4,7 +4,7 @@
     :module_author: Nathan Mendoza (nathancm@uci.edu)
 """
 
-from .tile_appearance import TileAppearance
+from .tile_appearance import TileAppearance, TileShape, TileColor
 from .movement_rule import MovementRule
 from .tile import Tile, NullTile
 from .tile_group import TileGroup
